Added Rebalance Tests(issues)
Added rebalance tests to oversized_bag_test but I ran into problems.  I attempted changing the input data and the clear function but was unable to fix the issue. The test file gets stuck in a loop when the rebalance tests are not commented out. This may be a problem with the rebalance function or the setup of the tests.