Update oversized_bag.hpp
Added load_existing_files function, a new constructor that takes file size as a parameter, and had each constructor run the new function to read relevant data. I was unable to verify the functionality as I could not figure out how to create files to be loaded. I tried using an output file to be read from build/test but I did not seem to read it. A different file location may work. New function does not cause a crash in my testing.